PAW-102 PAW-103 admin create animal with images

+++ Server/Repositories/AnimalRepository.cs
@@ -24,6 +24,12 @@ namespace PawsomeProject.Server.Repositories
             return await _dbContext.Animals.FindAsync(id);
         }
 
+        public async Task<List<Image>> GetImages(int id)
+        {
+            var images = await _dbContext.Images.Where(o => o.Animal.Id == id).ToListAsync();
+            return images;
+        }
+
         public async Task<Animal> AddAnimal(AnimalDto animalDto)
         {
             var newAnimal = new Animal
@@ -31,11 +37,27 @@ namespace PawsomeProject.Server.Repositories
                 Name = animalDto.Name,
                 Description = animalDto.Description,
                 Age = animalDto.Age,
-                Main_Image_Url = "",
+                Main_Image_Url = animalDto.Main_Image_Url,
+                Images = new List<Image>()
             };
-            var result = await _dbContext.Animals.AddAsync(newAnimal);
-            await _dbContext.SaveChangesAsync();
-            return result.Entity;
+            foreach (var image in animalDto.Images)
+            {
+                Image newImage = new Image
+                {
+                    URL = image.URL,
+                    Type = image.Type
+                };
+                newAnimal.Images.Add(newImage);
+            }
+
+            if (newAnimal != null)
+            {
+                var result = await _dbContext.Animals.AddAsync(newAnimal);
+                await _dbContext.SaveChangesAsync();
+                return result.Entity;
+            }
+
+            return null;
         }
 
         public async Task UpdateAnimal(AnimalDto animal)
@@ -47,8 +69,21 @@ namespace PawsomeProject.Server.Repositories
         public async Task DeleteAnimal(int id)
         {
             var animal = await _dbContext.Animals.FindAsync(id);
-            _dbContext.Animals.Remove(animal);
-            await _dbContext.SaveChangesAsync();
+            if (animal != null)
+            {
+                _dbContext.Animals.Remove(animal);
+                await _dbContext.SaveChangesAsync();
+            }
+        }
+
+        public async Task DeleteImage(int id)
+        {
+            var item = await _dbContext.Images.FindAsync(id);
+            if (item != null)
+            {
+                _dbContext.Images.Remove(item);
+                await _dbContext.SaveChangesAsync();
+            }
         }
     }
 }
